requirements.txt
streamlit
groq
python-docx
PyPDF2
sentence-transformers
chromadb
numpy
pandas
scikit-learn
duckdb
python-dotenv
requests
typing-extensions
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
langchain-groq
psycopg2-binary
